Exception-less Control Flow
Implemented an alternative control flow handling system which doesn't require exceptions. (It instead uses long/set_jmp). I anticipate this system will be less performant than the exception version (that's a scary statement) so right now it is disabled unless exceptions are disabled!

The interfaces of both systems are the same so the same code will work automatically either the presence or absence of exceptions!